Fixed part 1

+++ src/ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: jkaczmar <jkaczmar@student.42wolfsburg.de> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/17 11:55:57 by jkaczmar          #+#    #+#             */
-/*   Updated: 2021/11/22 15:09:34 by jkaczmar         ###   ########.fr       */
+/*   Updated: 2021/11/22 21:02:31 by jkaczmar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	int		counter;
+	size_t	counter;
 	char	*thingy;
 
 	counter = 0;
